requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
Babel==2.15.0
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
daphne==4.1.0
Django==5.0
django-ckeditor-5==0.2.13
django-extensions==3.2.3
django-health-check==3.18.3
django-js-asset==2.2.0
django-prometheus==2.3.1
django-redis==5.4.0
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
forex-python==1.8
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipaddress==1.0.23
kombu==5.4.0
lxml==5.2.1
MarkupSafe==2.1.5
midtransclient==1.4.2
msgpack==1.0.8
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
pkgconfig==1.5.5
premailer==3.10.0
priority==1.3.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
reportlab==4.2.0
requests==2.31.0
service-identity==24.1.0
setuptools==74.1.2
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.1
trio-websocket==0.11.1
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wsproto==1.2.0
xlwt==1.3.0

